Рефакторинг аннотаций возвращаемых типов, убраны лишние проверки

+++ repositories/address_repository.py
@@ -14,24 +14,18 @@ class AddressRepository:
         self.db.refresh(address)
         return address
 
-    def get_by_address_id(self, address_id: int) -> Address | bool:
+    def get_by_address_id(self, address_id: int) -> Address | None:
         address = self.db.get(Address, address_id)
-        if not address:
-            return False
         return address
 
-    def get_by_user_id(self, user_id: int) -> Address | bool:
+    def get_by_user_id(self, user_id: int) -> Address | None:
         query = select(Address).where(Address.user_id == user_id)
         result = self.db.execute(query).scalar_one_or_none()
-        if not result:
-            return False
         return result
 
-    def get_all_by_user_id(self, user_id) -> list[Address] | bool:
+    def get_all_by_user_id(self, user_id) -> list[Address] | None:
         query = select(Address).where(Address.user_id == user_id)
         result = self.db.execute(query).scalars().all()
-        if not result:
-            return False
         return result
 
     def update(self, address: list[Address]) -> list[Address]:
@@ -40,6 +34,8 @@ class AddressRepository:
 
     def delete(self, address_id: int) -> bool:
         address = self.get_by_address_id(address_id)
+        if address is None:
+            return False
         self.db.delete(address)
         self.db.flush()
         return True
